pipeline: Uses credentials provided in the git url
